requirements.txt
discord.py==2.1.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
SQLAlchemy==1.4.45
beautifulsoup4==4.11.1
requests==2.28.1
google-cloud-logging==3.3.1
google-cloud-secret-manager==2.13.0